requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.7.0
anyio==4.6.1
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==24.3.0
blinker==1.6.2
cachetools==5.3.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
deepdiff==6.5.0
dill==0.3.7
distlib==0.3.7
distro==1.9.0
exceptiongroup==1.1.3
filelock==3.12.3
fonttools==4.43.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.5.28
idna==3.7
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.5
jiter==0.6.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.8.0
numpy==1.25.2
openai==1.51.2
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
pillow==10.3.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.3.0
pre-commit==3.4.0
propcache==0.2.0
protobuf==4.24.3
pyarrow==14.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.0
Pygments==2.16.1
pylint==2.17.5
Pympler==1.0.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.32.3
rich==13.5.2
rpds-py==0.10.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
streamlit==1.39.0
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.7
validators==0.22.0
virtualenv==20.26.6
watchdog==3.0.0
wordcloud==1.9.2
wrapt==1.15.0
yarl==1.15.2
zipp==3.20.2
pytest-cov==4.1.0
